Make repo compatible with UNIX and Windows computers.
The .gitignore was added to ignore certain files that should be in the
same folder as the project files, but shouldn't be part of the repo. The
`.idea` and `sqlite.exe` files are good examples. Some of us might need
these files, but not all of us do. These two files were deleted from the
repo because of this.

`Fourguys.py` was modified to make sure python can find the database
file correctly on anyone's computer. The path given in the script for
the database URI was specific to a specific person's computer, when it
should work for all of us. Now, the database file should just be in the
same folder as `Fourguys.py`.

requirements.txt was added so that we have a list of all the packages
that are required for the project. If anyone installs new packages for
their project by using `pip`, then they should change the
requirements.txt file by doing the following:

1. run the command `pip freeze` in a console.
2. Take the output of the command and copy it into the requirements.txt
file.

If you're using the Bash shell, then you can just run `pip freeze >
requirements.txt`.

Changes to be committed:
      new file:   .gitignore
      deleted:    .idea/Fourguys.iml
      deleted:    .idea/inspectionProfiles/profiles_settings.xml
      deleted:    .idea/misc.xml
      deleted:    .idea/modules.xml
      modified:   Fourguys.py
      new file:   requirements.txt
      deleted:    sqlite3.exe