Place dummy data in csv files, remove database file.
The database file is not really a part of this project. If, when testing
the website, someone tests out the sign up functionality and adds a new
entry to the users database, the repository will have changed, because
the database file changed. However, the *project* didn't change. None of
the webpages changed, nothing important changed. So the database file is
something that each person should make, but not something that should be
shared for all of us.

Changes to be committed:
      deleted:    database.db
      new file:   dummy-data/menuitems.csv
      new file:   dummy-data/users.csv