Change server to use a manager, add extension `flask_script`.
Now that the database is separated from the repository, there should be
an easy and simple way to create the database from dummy data. Rather
than write new scripts to do that and only that, I changed Fourguys.py
to use the `flask_script` extension. This extension is mentioned in the
flask book that we all have. The basics are on page 37 of the pdf (page
17 of the book itself) under the section'Command-Line options with
Flask-Script'.

To rebuild the database from dummy data, run the following at a command
prompt, while inside of the virtual environemnt:

  python Fourguys.py rebuild_database

Here's some information on how flask script works:

  http://flask-script.readthedocs.io/en/latest/

It covers how to make new management functions.

Changes to be committed:
      modified:   Fourguys.py
      modified:   requirements.txt