Moved the static and templates directory inside the app directory
All of the HTML templates and static files that will be used
to render this web application are also moved into the new
app directory where the core of the application resides.

In addition, in the root of the project, we have created a
configuration file called config.py that includes all settings
that are needed to run the applications.

Lastly, the run.py is the script file that is going to be
executed to start the web application.